refactor: change package name to affine
