+++ src/main/java/module-info.java
@@ -1,5 +1,2 @@
-module ru.vsu.cs.konygina_d.render_engine {
-    requires transitive io.github.alphameo.linear_algebra;
-
-    exports ru.vsu.cs.konygina_d.render_engine;
+module $MODULE_NAME$ {
 }
