+++ src/main/java/ru/vsu/cs/konygina_d/affine/AffineTransformation.java
@@ -1,4 +1,4 @@
-package ru.vsu.cs.konygina_d.render_engine;
+package ru.vsu.cs.konygina_d.affine;
 
 import io.github.alphameo.linear_algebra.mat.Mat4Math;
 import io.github.alphameo.linear_algebra.mat.Matrix4;
