+++ src/main/java/ru/vsu/cs/konygina_d/affine/DataList.java
@@ -1,9 +1,13 @@
-package ru.vsu.cs.konygina_d.render_engine;
+package ru.vsu.cs.konygina_d.affine;
 
 public interface DataList<T> {
     void add(AffineTransformation at);
+
     void remove(int index);
+
     void remove(AffineTransformation at);
+
     void set(int index, AffineTransformation at);
+
     T get(int index);
 }
